:fire: [#65] remove contactpersoon and refactor to teams and OE
