+++ src/openorganisatie/scim/admin/__init__.py
@@ -1,5 +1,4 @@
 from .attr_mapping_config import AttribuutMappingConfigAdmin
-from .contactpersoon import ContactpersoonAdmin
 from .functie import FunctieAdmin
 from .functietype import FunctieTypeAdmin
 from .group import GroupAdmin
@@ -14,7 +13,6 @@ __all__ = [
     "OrganisatorischeEenheidAdmin",
     "VestigingAdmin",
     "TeamAdmin",
-    "ContactpersoonAdmin",
     "UserAdmin",
     "GroupAdmin",
     "FunctieAdmin",
