+++ src/openorganisatie/scim/api/serializers/organisatorische_eenheid.py
@@ -4,12 +4,30 @@ from rest_framework import serializers
 from vng_api_common.utils import get_help_text
 
 from openorganisatie.scim.models.functie import Functie
+from openorganisatie.scim.models.medewerker import Medewerker
 from openorganisatie.scim.models.organisatorische_eenheid import OrganisatorischeEenheid
 from openorganisatie.scim.models.vestiging import Vestiging
 from openorganisatie.utils.fields import UUIDRelatedField
 
 from ..serializers.functie import NestedFunctieSerializer
 from ..serializers.vestiging import VestigingSerializer
+
+
+class NestedMedewerkerSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Medewerker
+        fields = [
+            "uuid",
+            "medewerker_id",
+            "voornaam",
+            "achternaam",
+            "emailadres",
+            "telefoonnummer",
+            "geslachtsaanduiding",
+            "datum_uit_dienst",
+            "datum_toegevoegd",
+            "datum_aangepast",
+        ]
 
 
 class NestedOrganisatorischeEenheidSerializer(serializers.ModelSerializer):
@@ -32,11 +50,17 @@ class NestedOrganisatorischeEenheidSerializer(serializers.ModelSerializer):
             "emailadres",
             "telefoonnummer",
             "datum_opheffing",
+            "contactpersoon",
             "hoofd_organisatorische_eenheid",
         ]
 
 
 class OrganisatorischeEenheidSerializer(serializers.ModelSerializer):
+    contactpersoon = NestedMedewerkerSerializer(
+        read_only=True,
+        required=False,
+        help_text=get_help_text("scim.Team", "vestigingen"),
+    )
     vestigingen = VestigingSerializer(
         many=True,
         read_only=True,
@@ -82,6 +106,7 @@ class OrganisatorischeEenheidSerializer(serializers.ModelSerializer):
             "emailadres",
             "telefoonnummer",
             "datum_opheffing",
+            "contactpersoon",
             "vestigingen",
             "vestigingen_uuid",
             "functies",
