+++ src/openorganisatie/scim/api/urls.py
@@ -7,7 +7,6 @@ from vng_api_common import routers
 
 from ...utils.views import SpectacularJSONAPIView, SpectacularYAMLAPIView
 from .schema import custom_settings
-from .viewsets.contactpersoon import ContactpersoonViewSet
 from .viewsets.functie import FunctieViewSet
 from .viewsets.functietype import FunctieTypeViewSet
 from .viewsets.group import GroupReadOnlyViewSet
@@ -21,7 +20,6 @@ app_name = "scim_api"
 
 router = routers.DefaultRouter()
 router.register("medewerkers", MedewerkerViewSet)
-router.register("contactpersoon", ContactpersoonViewSet)
 router.register("vestiging", VestigingViewSet)
 router.register("organisatorische-eenheid", OrganisatorischeEenheidViewSet)
 router.register("teams", TeamViewSet)
