+++ src/openorganisatie/scim/models/__init__.py
@@ -1,5 +1,4 @@
 from .attr_mapping_config import AttribuutMappingConfig
-from .contactpersoon import Contactpersoon
 from .functie import Functie
 from .functietype import FunctieType
 from .medewerker import Medewerker
@@ -7,7 +6,6 @@ from .vestiging import Vestiging
 
 __all__ = [
     "Medewerker",
-    "Contactpersoon",
     "Vestiging",
     "Functie",
     "FunctieType",
